fix: timeline selection
