+++ screenpipe-app-tauri/components/timeline/timeline-block.tsx
@@ -1,5 +1,7 @@
 import { StreamTimeSeriesResponse } from "@/app/timeline/page";
-import { useMemo } from "react";
+import { useTimelineSelection } from "@/lib/hooks/use-timeline-selection";
+import { debounce } from "lodash";
+import { useMemo, useState, useRef, useEffect } from "react";
 
 interface TimeBlock {
   appName: string;
@@ -14,59 +16,80 @@ interface TimelineBlocksProps {
   timeRange: {
     start: Date;
     end: Date;
-    visibleStart?: Date;
-    visibleEnd?: Date;
   };
 }
 
-export function TimelineBlocks({ frames, timeRange }: TimelineBlocksProps) {
-  // Cache colors to avoid recalculating
-  const colorCache = useMemo(() => new Map<string, string>(), []);
+// First, memoize the color generation function
+const useColorGenerator = () => {
+  return useMemo(() => {
+    const colorCache = new Map<string, string>();
 
-  // Calculate blocks without sampling
-  const blocks = useMemo(() => {
-    const getAppColor = (appName: string): string => {
+    return (appName: string): string => {
       const cached = colorCache.get(appName);
       if (cached) return cached;
 
-      // Use a better hash distribution
       const hash = Array.from(appName).reduce(
         (h, c) => (Math.imul(31, h) + c.charCodeAt(0)) | 0,
         0
       );
 
-      // Use golden ratio with a larger step for more distinct hues
       const golden_ratio = 0.618033988749895;
       const hue = ((hash * golden_ratio * 1.5) % 1) * 360;
-
-      // Increase saturation and use fixed lightness for better distinction
-      const sat = 85 + (hash % 15); // 85-100%
-      const light = 60; // Fixed lightness for better visibility
+      const sat = 85 + (hash % 15);
+      const light = 60;
 
       const color = `hsl(${hue}, ${sat}%, ${light}%)`;
       colorCache.set(appName, color);
       return color;
     };
+  }, []); // Empty deps since this never needs to change
+};
+
+export function TimelineBlocks({ frames, timeRange }: TimelineBlocksProps) {
+  const getAppColor = useColorGenerator();
+  const { setSelectionRange } = useTimelineSelection();
+
+  // Memoize the sorted frames
+  const sortedFrames = useMemo(
+    // TODO: this should not be necessary
+    () =>
+      [...frames].sort(
+        (a, b) =>
+          new Date(a.timestamp).getTime() - new Date(b.timestamp).getTime()
+      ),
+    [frames]
+  );
+
+  // Memoize the block generation with precise dependencies
+  const blocks = useMemo(() => {
     if (frames.length === 0) return [];
 
-    const blocks: TimeBlock[] = [];
+    const visibleBlocks: (TimeBlock & { isGap?: boolean })[] = [];
     let currentBlock: TimeBlock | null = null;
+    let lastEndTime: Date | null = null;
 
-    // Sort frames by timestamp first
-    const sortedFrames = [...frames].sort(
-      (a, b) =>
-        new Date(a.timestamp).getTime() - new Date(b.timestamp).getTime()
-    );
-
-    sortedFrames.forEach((frame) => {
+    for (const frame of sortedFrames) {
       const timestamp = new Date(frame.timestamp);
-      // Ensure we have devices data
-      if (!frame.devices?.[0]?.metadata?.app_name) return;
+      if (!frame.devices?.[0]?.metadata?.app_name) continue;
 
       const appName = frame.devices[0].metadata.app_name;
       const windowName = frame.devices[0].metadata.window_name;
-      if (timestamp < timeRange.start || timestamp > timeRange.end) return;
 
+      if (timestamp < timeRange.start || timestamp > timeRange.end) continue;
+
+      // Handle gap detection
+      if (lastEndTime && timestamp.getTime() - lastEndTime.getTime() > 1000) {
+        visibleBlocks.push({
+          appName: "gap",
+          windowName: "No Activity",
+          startTime: lastEndTime,
+          endTime: timestamp,
+          color: "transparent",
+          isGap: true,
+        });
+      }
+
+      // Handle block creation/update
       if (!currentBlock) {
         currentBlock = {
           appName,
@@ -76,7 +99,7 @@ export function TimelineBlocks({ frames, timeRange }: TimelineBlocksProps) {
           color: getAppColor(appName),
         };
       } else if (currentBlock.appName !== appName) {
-        blocks.push(currentBlock);
+        visibleBlocks.push(currentBlock);
         currentBlock = {
           appName,
           windowName,
@@ -87,41 +110,143 @@ export function TimelineBlocks({ frames, timeRange }: TimelineBlocksProps) {
       } else {
         currentBlock.endTime = timestamp;
       }
-    });
 
-    if (currentBlock) blocks.push(currentBlock);
-    return blocks;
-  }, [frames, timeRange]);
+      lastEndTime = timestamp;
+    }
+
+    if (currentBlock) visibleBlocks.push(currentBlock);
+
+    // Handle edge gaps
+    if (visibleBlocks[0]?.startTime.getTime() > timeRange.start.getTime()) {
+      visibleBlocks.unshift({
+        appName: "gap",
+        windowName: "No Activity",
+        startTime: timeRange.start,
+        endTime: visibleBlocks[0].startTime,
+        color: "transparent",
+        isGap: true,
+      });
+    }
+
+    const lastBlock = visibleBlocks[visibleBlocks.length - 1];
+    if (lastBlock && lastBlock.endTime.getTime() < timeRange.end.getTime()) {
+      visibleBlocks.push({
+        appName: "gap",
+        windowName: "No Activity",
+        startTime: lastBlock.endTime,
+        endTime: timeRange.end,
+        color: "transparent",
+        isGap: true,
+      });
+    }
+
+    return visibleBlocks;
+  }, [
+    sortedFrames, // Use sorted frames instead of raw frames
+    timeRange.start, // Use timestamp instead of Date object
+    timeRange.end,
+    getAppColor,
+  ]);
+
+  // Add selection state
+  const [isDragging, setIsDragging] = useState(false);
+  const [selectionStart, setSelectionStart] = useState<number | null>(null);
+  const [selectionEnd, setSelectionEnd] = useState<number | null>(null);
+  const containerRef = useRef<HTMLDivElement>(null);
+
+  // Add mouse handlers
+  const handleMouseDown = (e: React.MouseEvent) => {
+    const element = document.elementFromPoint(e.clientX, e.clientY);
+    const blockIndex = element?.getAttribute("data-block-index");
+    if (blockIndex) {
+      const index = parseInt(blockIndex);
+      setSelectionStart(index);
+      setSelectionEnd(index);
+      setSelectionRange({
+        start: blocks[index].startTime,
+        end: blocks[index].endTime,
+      });
+      setIsDragging(true);
+    }
+  };
+
+  // Add debouncing to mouse move handler
+  const handleMouseMove = useMemo(
+    () =>
+      debounce((e: React.MouseEvent) => {
+        if (!isDragging || selectionStart === null) return;
+        const element = document.elementFromPoint(e.clientX, e.clientY);
+        const blockIndex = element?.getAttribute("data-block-index");
+        if (blockIndex) {
+          setSelectionEnd(parseInt(blockIndex));
+          setSelectionRange({
+            start: blocks[selectionStart].startTime,
+            end: blocks[parseInt(blockIndex)].endTime,
+          });
+        }
+      }, 16), // ~60fps
+    [isDragging, selectionStart]
+  );
+
+  const handleMouseUp = () => {
+    setIsDragging(false);
+  };
+
+  // Calculate selection bounds only when needed
+  const selectionBounds = useMemo(() => {
+    if (selectionStart === null || selectionEnd === null) return null;
+
+    const start = Math.min(selectionStart, selectionEnd);
+    const end = Math.max(selectionStart, selectionEnd);
+
+    return {
+      left:
+        ((blocks[start].startTime.getTime() - timeRange.start.getTime()) /
+          (timeRange.end.getTime() - timeRange.start.getTime())) *
+        100,
+      width:
+        ((blocks[end].endTime.getTime() - blocks[start].startTime.getTime()) /
+          (timeRange.end.getTime() - timeRange.start.getTime())) *
+        100,
+    };
+  }, [selectionStart, selectionEnd, blocks, timeRange]);
 
   return (
-    <div className="absolute inset-0 flex flex-col">
+    <div
+      ref={containerRef}
+      className="absolute inset-0 flex flex-col"
+      onMouseDown={handleMouseDown}
+      onMouseUp={handleMouseUp}
+      onMouseMove={handleMouseMove}
+      style={{ userSelect: "none" }}
+    >
+      {/* Regular blocks */}
       {blocks.map((block, index) => {
-        // Calculate position relative to visible range instead of 24-hour scale
-        const visibleStartTime = timeRange.visibleStart || timeRange.start;
-        const visibleEndTime = timeRange.visibleEnd || timeRange.end;
-        const visibleRangeMs =
-          visibleEndTime.getTime() - visibleStartTime.getTime();
-
-        const blockStartMs =
-          block.startTime.getTime() - visibleStartTime.getTime();
-        const blockDurationMs =
-          block.endTime.getTime() - block.startTime.getTime();
-
-        // Calculate percentages based on visible range
-        const blockStart = (blockStartMs / visibleRangeMs) * 100;
-        const blockWidth = (blockDurationMs / visibleRangeMs) * 100;
-
-        if (blockWidth < 0.01) return null; // Skip tiny blocks
+        const blockStart =
+          ((block.startTime.getTime() - timeRange.start.getTime()) /
+            (timeRange.end.getTime() - timeRange.start.getTime())) *
+          100;
+        const blockWidth =
+          ((block.endTime.getTime() - block.startTime.getTime()) /
+            (timeRange.end.getTime() - timeRange.start.getTime())) *
+          100;
 
         return (
           <div
             key={`${block.appName}-${index}`}
-            className="absolute top-0 bottom-0 opacity-50 hover:opacity-80 transition-opacity z-10"
+            data-block-index={index}
+            className={`absolute top-0 bottom-0 transition-opacity z-10 ${
+              block.isGap
+                ? "hover:bg-gray-200 dark:hover:bg-gray-700 opacity-0 hover:opacity-30"
+                : "opacity-50 hover:opacity-80"
+            }`}
             style={{
-              left: `${blockStart}%`,
-              width: `${blockWidth}%`,
+              transform: `translateX(${blockStart}%) scaleX(${
+                blockWidth / 100
+              })`,
+              transformOrigin: "left",
+              width: "100%",
               backgroundColor: block.color,
-              willChange: "transform",
             }}
             title={`${block.appName}\n${
               block.windowName
@@ -129,6 +254,18 @@ export function TimelineBlocks({ frames, timeRange }: TimelineBlocksProps) {
           />
         );
       })}
+
+      {/* Selection overlay */}
+      {selectionBounds && (
+        <div
+          className="absolute top-0 bottom-0 bg-black/20 dark:bg-white/20 z-20"
+          style={{
+            left: `${selectionBounds.left}%`,
+            width: `${selectionBounds.width}%`,
+            willChange: "transform",
+          }}
+        />
+      )}
     </div>
   );
 }
