+++ screenpipe-app-tauri/components/timeline/timeline-dock.tsx
@@ -320,7 +320,7 @@ export function TimelineIconsSection({
   }, [processedBlocks, loadAppIcon]);
 
   return (
-    <div className="absolute -top-12 inset-x-0 pointer-events-none h-8">
+    <div className="absolute -top-8 inset-x-0 pointer-events-none h-8">
       {processedBlocks.map((block, i) => (
         <div
           key={`${block.appName}-${i}`}
